Update ProjectFileParser.groovy
Fixed a bug with the building of WinPhone project. On 64 bit systems, the default variable MSBuildExtensionsPath=MSBuildExtensionsPath64. However, the default installation path of WinPhone sdk contains MSBuildExtensionsPath32. But some of the files needed to build contains the path MSBuildExtensionsPath64.So now plug-in trying to find the imported files both ways: having MSBuildExtensionsPath32 and MSBuildExtensionsPath64.