Use Evaluation.Project to parse project file
Instead of using the deprecated Project class that is not compatible
with the msbuild 4.0+ format, we use the non deprecated to allow
parsing of all the new things introduce in the format like the Label
on Import tag.

The parsing of the solution is done by internal class: SolutionParser
A wrapper over all concerned classes has been created.

The serialization of the Solution still use the old method because the
new parser don't give you all the information like the buildOrder.